+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/DockerConnector.java
@@ -7,9 +7,10 @@ import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.DockerClie
 import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.exception.DockerException;
 import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.Version;
 import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.DefaultDockerClientConfig;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.DockerClientConfig;
+import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.RemoteApiVersion;
 import com.github.kostyasha.yad.docker_java.com.google.common.base.Preconditions;
 import com.github.kostyasha.yad.docker_java.org.apache.commons.lang.StringUtils;
+import com.github.kostyasha.yad.other.ConnectorType;
 import com.github.kostyasha.yad.utils.CredentialsListBoxModel;
 import com.google.common.base.Throwables;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
@@ -36,6 +37,12 @@ import java.util.Collections;
 import java.util.List;
 
 import static com.github.kostyasha.yad.client.ClientBuilderForConnector.newClientBuilderForConnector;
+import static com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.RemoteApiVersion.parseConfig;
+import static com.github.kostyasha.yad.other.ConnectorType.NETTY;
+import static hudson.util.FormValidation.ok;
+import static hudson.util.FormValidation.warning;
+import static org.apache.commons.lang.builder.ToStringBuilder.reflectionToString;
+import static org.apache.commons.lang.builder.ToStringStyle.MULTI_LINE_STYLE;
 
 /**
  * Settings for connecting to docker.
@@ -48,15 +55,17 @@ public class DockerConnector implements Describable<DockerConnector> {
     private String serverUrl;
 
     @CheckForNull
-    private String apiVersion = null;
+    private String apiVersion;
 
-    private Boolean tlsVerify = true;
+    private transient Boolean tlsVerify;
 
     @CheckForNull
     private String credentialsId = null;
 
     @CheckForNull
     private transient DockerClient client = null;
+
+    private ConnectorType connectorType = NETTY;
 
     @DataBoundConstructor
     public DockerConnector(String serverUrl) {
@@ -82,15 +91,6 @@ public class DockerConnector implements Describable<DockerConnector> {
         this.apiVersion = StringUtils.trimToNull(apiVersion);
     }
 
-    public Boolean getTlsVerify() {
-        return tlsVerify;
-    }
-
-    @DataBoundSetter
-    public void setTlsVerify(Boolean tlsVerify) {
-        this.tlsVerify = tlsVerify;
-    }
-
     public String getCredentialsId() {
         return credentialsId;
     }
@@ -98,6 +98,19 @@ public class DockerConnector implements Describable<DockerConnector> {
     @DataBoundSetter
     public void setCredentialsId(String credentialsId) {
         this.credentialsId = credentialsId;
+    }
+
+    /**
+     * @see #connectorType
+     */
+    @CheckForNull
+    public ConnectorType getConnectorType() {
+        return connectorType;
+    }
+
+    @DataBoundSetter
+    public void setConnectorType(ConnectorType connectorType) {
+        this.connectorType = connectorType;
     }
 
     public DockerClient getClient() {
@@ -144,6 +157,8 @@ public class DockerConnector implements Describable<DockerConnector> {
                 .append(serverUrl, that.serverUrl)
                 .append(apiVersion, that.apiVersion)
                 .append(credentialsId, that.credentialsId)
+                .append(tlsVerify, that.tlsVerify)
+                .append(connectorType, that.connectorType)
                 .isEquals();
     }
 
@@ -153,6 +168,8 @@ public class DockerConnector implements Describable<DockerConnector> {
                 .append(serverUrl)
                 .append(apiVersion)
                 .append(credentialsId)
+                .append(tlsVerify)
+                .append(connectorType)
                 .toHashCode();
     }
 
@@ -179,28 +196,45 @@ public class DockerConnector implements Describable<DockerConnector> {
         public FormValidation doTestConnection(
                 @QueryParameter String serverUrl,
                 @QueryParameter String apiVersion,
-                @QueryParameter Boolean tlsVerify,
-                @QueryParameter String credentialsId
+                @QueryParameter String credentialsId,
+                @QueryParameter ConnectorType connectorType
         ) throws IOException, ServletException, DockerException {
             try {
-                final DockerClientConfig clientConfig = new DefaultDockerClientConfig.Builder()
+                DefaultDockerClientConfig.Builder configBuilder = new DefaultDockerClientConfig.Builder()
                         .withApiVersion(apiVersion)
-                        .withDockerHost(serverUrl)
-                        .withDockerTlsVerify(tlsVerify)
-                        .build();
-
+                        .withDockerHost(serverUrl);
 
                 final DockerClient testClient = newClientBuilderForConnector()
-                        .withDockerClientConfig(clientConfig)
+                        .withConfigBuilder(configBuilder)
+                        .withConnectorType(connectorType)
                         .withCredentials(credentialsId)
                         .build();
 
                 Version verResult = testClient.versionCmd().exec();
 
-                return FormValidation.ok(verResult.toString());
+                return ok(reflectionToString(verResult, MULTI_LINE_STYLE));
             } catch (Exception e) {
                 return FormValidation.error(e, e.getMessage());
             }
+        }
+
+        public FormValidation doCheckApiVersion(@QueryParameter String apiVersion) {
+            if (StringUtils.isEmpty(apiVersion)) {
+                return ok();
+            }
+            try {
+                final RemoteApiVersion rav = parseConfig(apiVersion);
+//                if (rav.isGreaterOrEqual(VERSION_1_24)) {
+//                    return warning("Latest tested version 1.24. Current configuration may not work correctly");
+//                } else
+                if (!rav.isGreaterOrEqual(RemoteApiVersion.VERSION_1_19)) {
+                    return warning("Unknown API version, may not work with plugin!");
+                }
+            } catch (Exception ex) {
+                return FormValidation.error("Can't parse api version.", ex);
+            }
+
+            return ok();
         }
 
         @Override