+++ yet-another-docker-plugin/src/main/resources/com/github/kostyasha/yad/DockerConnector/config.groovy
@@ -19,15 +19,17 @@ f.entry(title: _("Docker API version"), field: "apiVersion") {
     f.textbox()
 }
 
-f.entry(title: _("TLS verify"), field: "tlsVerify") {
-    f.checkbox()
-}
-
 f.entry(title: _("Host Credentials"), field: "credentialsId") {
     c.select()
 }
 
+f.entry(title: "Type", field: "connectorType") {
+    f.enum() {
+        text(my.name())
+    }
+}
+
 f.validateButton(title: _("Test Connection"), progress: _("Testing..."),
         method: "testConnection",
-        with: "serverUrl,credentialsId,version,connectTimeout,readTimeout"
+        with: "serverUrl,credentialsId,version,connectorType"
 )